extractor/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
azure-storage-file-datalake
